Add KAU LDD project for temperature sensor and LED blinking device
This adds a project that implements the KAU LDD temperature sensor and LED
blinking device. The project is based on the BulkVendor demo, but changed to
report temperature and support changing the LED state.

The device presents itself with vendor ID 03eb and device ID 206c and contains
to interrupt type endpoints for input and output. The input endpoint sends
simple messages each consisting of two bytes. The first byte is a signed integer
that represents the current temperature reading in degrees celcius. The second
byte is an unsigned integer with a bitmask of the currently enabled LEDs (the
lower four bits represent the four LEDs available on the device).

The output endpoint will read single byte messages from the host which are
simply applied as a new bitmap of enabled LEDs.

The project also contains a Python script that will communicate with the device.
This script has two operating modes: One that uses the pyusb library to talk
directly to the device, and one that opens the device file (/dev/kauldd0)
exported by the kernel mode driver. The script will run in libusb mode by
default, and switch to kernel mode if passed 'kernel' as a command line
parameter.

Signed-off-by: Toke Høiland-Jørgensen <toke.hoiland-jorgensen@kau.se>